[nvim] Add spellchecking!
Use \s to toggle spellchecking for a file.
Use Ctrl-L in insert mode to fix the most recent spelling error.
Use [s and ]s to jump between spelling errors in normal mode
Use z= to see a list of corrections.
Use zg to mark a word as globally good and zG for the current buffer